fix: rename notifcation → notification (typo in filenames and imports)
Renomme types/notifcation.model.ts et src/actions/notifcation.action.ts
vers leurs noms corrects, et met à jour les 8 fichiers qui les importaient.
