+++ app/(protected)/notification/[id]/page.tsx
@@ -1,13 +1,13 @@
 import { Suspense } from 'react';
 import Loading from '@/components/layouts/loading';
 import { DetailNotification } from './content';
-import { fetchDetailNotifcation } from '@/src/actions/notifcation.action';
+import { fetchDetailNotifcation } from '@/src/actions/notification.action';
 
 export default async function Page({ params }: { params: { id: string } }) {
-    const notifcationDetail = await fetchDetailNotifcation(params.id ?? '');
+    const notificationDetail = await fetchDetailNotifcation(params.id ?? '');
     return (
         <Suspense fallback={<Loading />}>
-            <DetailNotification detailNotification={notifcationDetail} />
+            <DetailNotification detailNotification={notificationDetail} />
         </Suspense>
     );
 }
