+++ app/(protected)/notification/controller.tsx
@@ -1,7 +1,7 @@
 
 
-import { fetchAllNotifcation } from "@/src/actions/notifcation.action";
-import { NotificationVM } from "@/types/notifcation.model";
+import { fetchAllNotifcation } from "@/src/actions/notification.action";
+import { NotificationVM } from "@/types/notification.model";
 import { useSession } from "next-auth/react";
 import { useEffect, useState } from "react";
