+++ types/notification.model.ts
